hpm_sdk release v0.13.1
Main changes since 0.13.0

- Changed:
  - Driver:I2S: update i2s drivers
  - middleware: lwip: optimize variable_name definition
  - samples: lwip: lwip_iperf: update readme files
  - samples: lwip: common: arch: add LWIP_MEM_SECTION declartion

- Fixed:
  - samples: multicore: BOOT_HEADER was missing in multicore core0 example
  - samples: jpeg: Fix encoding and decoding problem
  - samples: audio codec: wav decoder: fix 32bit wave file playback
  - I2S_DMA: fix wav channel not align problem
  - i2s_interrupt: fix I2S FIFO overflow
  - fix lack of interrupt claim for swi
  - driver: watchdog: overflow
  - Fix critical section logic issue in dma manager
  - Fix the core1 application debugging issue

+++ drivers/src/hpm_wdg_drv.c
@@ -124,7 +124,7 @@ interrupt_interval_t wdg_convert_interrupt_interval_from_us(const uint32_t src_f
         src_clk_one_tick_in_ns = 1U;
     }
 
-    uint32_t interrupt_interval_ticks = (uint32_t) ((uint64_t) interval_us * 1000UL) / src_clk_one_tick_in_ns;
+    uint32_t interrupt_interval_ticks = ((uint64_t) interval_us * 1000UL) / src_clk_one_tick_in_ns;
 
     for (uint32_t i = 0; i < ARRAY_SIZE(k_interrupt_interval_map); i++) {
         if (interrupt_interval_ticks <= k_interrupt_interval_map[i].top) {
